feat: Update HeroItem component
- 새로운 디자인을 적용하였습니다
- size prop을 추가하였습니다
- 리스트 바깥에서도 쓰일 수 있게 태그를 수정하였습니다
- 기능 관련한 코드와 태그를 삭제하고 시각적인 요소만 가지도록 수정하였습니다
