+++ src/components/MissionList/HeroItem/HeroItem.module.css
@@ -1,0 +1,85 @@
+.li {
+  list-style: none;
+  position: relative;
+  padding: var(--spacing-xs);
+  border-radius: var(--border-radius-sm);
+  box-shadow: var(--shadow-default);
+  text-align: center;
+
+  &.sm {
+    width: 160px;
+    height: 160px;
+  }
+
+  &.lg {
+    width: 300px;
+    height: 300px;
+  }
+
+  &.main {
+    background: #fff linear-gradient(180deg, #6885eb 1.86%, #fff 98.23%);
+  }
+  &.black {
+    background: #fff linear-gradient(180deg, #6885eb 1.86%, #fff 98.23%);
+    background: #fff
+      linear-gradient(180deg, var(--grayscale-darkgray) 1.86%, #fff 98.23%);
+    filter: drop-shadow(1px 3px 4px rgba(0, 0, 0, 0.25));
+  }
+}
+
+.li.receiver::after {
+  cursor: not-allowed;
+  position: absolute;
+  top: 75%;
+  left: 75%;
+  content: '';
+  width: 75%;
+  height: 75%;
+
+  background-image: url('/svgs/check.svg');
+  background-size: 75%;
+  background-repeat: no-repeat;
+  border-radius: var(--border-radius-sm);
+  transform: translate3d(-75%, -75%, 0);
+}
+
+.heroButton {
+  cursor: pointer;
+  width: 100%;
+  height: 100%;
+  padding: var(--spacing-xs);
+  background: #fff linear-gradient(180deg, #6885eb 1.86%, #fff 98.23%);
+  border-radius: var(--border-radius-sm);
+  box-shadow: var(--shadow-default);
+}
+
+.li.selected .heroButton {
+  border: 4px solid var(--color-main-blue);
+}
+.li.receiver .heroButton {
+  cursor: not-allowed;
+  background: #fff
+    linear-gradient(180deg, var(--grayscale-darkgray) 1.86%, #fff 98.23%);
+  filter: drop-shadow(1px 3px 4px rgba(0, 0, 0, 0.25));
+}
+
+.profileImage {
+  width: 100%;
+  height: 80%;
+  object-fit: contain;
+  border-radius: var(--border-radius-sm);
+}
+
+.heroName {
+  font-family: var(--font-family-content);
+  font-weight: var(--text-weight-bold);
+  text-shadow: var(--shadow-default);
+  color: var(--grayscale-black);
+
+  &.sm {
+    font-size: var(--text-size-base);
+  }
+  &.lg {
+    font-size: var(--text-size-subTitle);
+  }
+}
