+++ src/components/MissionList/HeroItem/index.ts
@@ -1,0 +1,1 @@
+export * from './HeroItem';
