--- src/components/MissionList/HeroList/HeroItem.module.css
@@ -1,53 +1,0 @@
-.li {
-  list-style: none;
-  position: relative;
-  width: 90px;
-  height: 94px;
-}
-
-.li.receiver::after {
-  cursor: not-allowed;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  content: '';
-  width: 50px;
-  height: 50px;
-
-  background-image: url('/svgs/check.svg');
-  background-size: 50px;
-  background-repeat: no-repeat;
-  border-radius: var(--border-radius-sm);
-  transform: translate3d(-50%, -50%, 0);
-}
-
-.heroButton {
-  cursor: pointer;
-  width: 100%;
-  height: 100%;
-  padding: var(--spacing-xs);
-  background: #fff1a9 linear-gradient(180deg, #ea8274 1.86%, #fff1a9 98.23%);
-  border-radius: var(--border-radius-sm);
-  box-shadow: var(--shadow-default);
-}
-.li.selected .heroButton {
-  border: 4px solid var(--color-green);
-}
-.li.receiver .heroButton {
-  cursor: not-allowed;
-  filter: grayscale(100%);
-}
-
-.profileImage {
-  width: 100%;
-  height: 80%;
-  object-fit: contain;
-  border-radius: var(--border-radius-sm);
-}
-
-.heroName {
-  color: var(--grayscale-black);
-  font-family: var(--font-family-content);
-  font-weight: var(--text-weight-bold);
-  text-shadow: var(--shadow-default);
-}
