+++ src/styles/custom-base.css
@@ -1,5 +1,5 @@
 *:focus {
-  outline: 5px solid var(--color-green);
+  outline: 5px solid var(--color-main-blue);
 }
 
 body {
